+++ matchbox-server/src/main/java/ch/ahdis/matchbox/terminology/TerminologyUtils.java
@@ -29,8 +29,7 @@ public class TerminologyUtils {
 		return parameters;
 	}
 
-	public static Parameters createSuccessfulResponseParameters(@Nullable final Coding coding,
-																					@Nullable final CodeableConcept codeableConcept) {
+	public static Parameters createSuccessfulResponseParameters(final Coding coding) {
 		final var parameters = new Parameters();
 		parameters.setParameter("result", true);
 		if (coding.hasVersion()) {
@@ -45,28 +44,33 @@ public class TerminologyUtils {
 		if (coding.hasDisplay()) {
 			parameters.setParameter("display", coding.getDisplayElement());
 		}
-		if (codeableConcept != null) {
-			parameters.setParameter("codeableConcept", codeableConcept);
-		}
 		return parameters;
 	}
 
-	public static Parameters mapCodeErrorToParameters(final String message,
-																	  final Coding coding) {
+	public static Parameters createErrorResponseParameters(final String message,
+																			 @Nullable final Coding coding,
+																			 @Nullable final CodeableConcept codeableConcept) {
 		final var parameters = new Parameters();
 		parameters.setParameter("result", false);
 		parameters.setParameter("message", message);
-		if (coding.hasVersion()) {
-			parameters.setParameter("version", coding.getVersionElement());
+
+		if (coding != null) {
+			if (coding.hasVersion()) {
+				parameters.setParameter("version", coding.getVersionElement());
+			}
+			if (coding.hasCode()) {
+				parameters.setParameter("code", coding.getCodeElement());
+			}
+			if (coding.hasSystem()) {
+				parameters.setParameter("system", coding.getSystemElement());
+			}
+			if (coding.hasDisplay()) {
+				parameters.setParameter("display", coding.getDisplayElement());
+			}
 		}
-		if (coding.hasCode()) {
-			parameters.setParameter("code", coding.getCodeElement());
-		}
-		if (coding.hasSystem()) {
-			parameters.setParameter("system", coding.getSystemElement());
-		}
-		if (coding.hasDisplay()) {
-			parameters.setParameter("display", coding.getDisplayElement());
+
+		if (codeableConcept != null) {
+			parameters.setParameter("codeableConcept", codeableConcept);
 		}
 
 		final var oo = new OperationOutcome();