+++ docs/changelog.md
@@ -2,6 +2,7 @@ Unreleased
 - Tutorial for validation FHIR resources with [matchbox](https://ahdis.github.io/matchbox/validation-tutorial/)
 - Gazelle reports: add test to ensure https://gazelle.ihe.net/jira/browse/EHS-831 is fixed
 - Allow validating a resource through the GUI with URL search parameters [#288](https://github.com/ahdis/matchbox/issues/288)
+- Terminology: support CodeableConcept in ValueSet/$validate operation [#291](https://github.com/ahdis/matchbox/issues/291)
 
 2024/10/07 Release 3.9.3